Revert to a stable state.
This is not expected to pass tests, since it provokes some known bugs whose
fixes have been or will be submitted in parallel.

There's a bug in makeQueued in liveSlots affecting promises.
And there's flakey string compare in the 'contractHost Demo --trivial' test.

I also re-instated changes to convert 'invite' to 'invitePayment' Those should be transparent.